+++ relays/ethereum/src/exchange.rs
@@ -54,7 +54,7 @@ pub trait SourceBlock {
 	type Transaction: SourceTransaction;
 
 	/// Return hash of the block.
-	fn id(&self) -> crate::sync_types::HeaderId<Self::Hash, Self::Number>;
+	fn id(&self) -> crate::utils::HeaderId<Self::Hash, Self::Number>;
 	/// Return block transactions iterator.
 	fn transactions(&self) -> Vec<Self::Transaction>;
 }
@@ -81,7 +81,7 @@ pub type TransactionOf<P> = <<P as TransactionProofPipeline>::Block as SourceBlo
 pub type TransactionHashOf<P> = <TransactionOf<P> as SourceTransaction>::Hash;
 
 /// Header id.
-pub type HeaderId<P> = crate::sync_types::HeaderId<BlockHashOf<P>, BlockNumberOf<P>>;
+pub type HeaderId<P> = crate::utils::HeaderId<BlockHashOf<P>, BlockNumberOf<P>>;
 
 /// Source client API.
 #[async_trait]
@@ -443,7 +443,7 @@ async fn wait_header_finalized<P: TransactionProofPipeline>(
 #[cfg(test)]
 pub(crate) mod tests {
 	use super::*;
-	use crate::sync_types::HeaderId;
+	use crate::utils::HeaderId;
 
 	use parking_lot::Mutex;
 	use std::{
