+++ relays/ethereum/src/sync_loop_tests.rs
@@ -17,8 +17,8 @@
 #![cfg(test)]
 
 use crate::sync_loop::{run, SourceClient, TargetClient};
-use crate::sync_types::{HeaderId, HeadersSyncPipeline, QueuedHeader, SourceHeader, SubmittedHeaders};
-use crate::utils::{process_future_result, retry_backoff, MaybeConnectionError};
+use crate::sync_types::{HeadersSyncPipeline, QueuedHeader, SourceHeader, SubmittedHeaders};
+use crate::utils::{process_future_result, retry_backoff, HeaderId, MaybeConnectionError};
 
 use async_trait::async_trait;
 use backoff::backoff::Backoff;
